requirements.txt
fastapi
uvicorn
numpy
joblib
python-dotenv
